Changed a sentence
Changed a sentence to stay more with the english original translation. 
Before: helps terminating everything
After: helps terminating everything and everyone whats preventing the mission to be a success. (kinda that)

Also, changed the 939 sentence to : sense movements not see movements. Which is basicly right but isnt fitting the english translation also.